+++ esp32c3-embassy/src/random.rs
@@ -1,13 +1,14 @@
-// Copyright Claudio Mattera 2024.
+// Copyright Claudio Mattera 2024-2025.
 //
-// This Source Code Form is subject to the terms of the Mozilla Public
-// License, v. 2.0. If a copy of the MPL was not distributed with this
-// file, You can obtain one at http://mozilla.org/MPL/2.0/.
+// Distributed under the MIT License or the Apache 2.0 License at your option.
+// See the accompanying files LICENSE-MIT.txt and LICENSE-APACHE-2.0.txt, or
+// online at
+// https://opensource.org/licenses/MIT
+// https://opensource.org/licenses/Apache-2.0
 
 //! Random numbers generator
 
 use rand_core::CryptoRng;
-use rand_core::Error;
 use rand_core::RngCore;
 
 use esp_hal::rng::Rng;
@@ -37,11 +38,6 @@ impl RngCore for RngWrapper {
             let [random_value, _, _, _] = self.next_u32().to_ne_bytes();
             *value = random_value;
         }
-    }
-
-    fn try_fill_bytes(&mut self, dest: &mut [u8]) -> Result<(), Error> {
-        self.fill_bytes(dest);
-        Ok(())
     }
 }