+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0] - 2025-02-07
+
+### Changed
+
+- Use [Adafruit IO](https://io.adafruit.com) API instead of WorldTimeAPI to fetch current time
+- Change license to MIT + APACHE-2.0 for main firmware
+- Update `embassy-executor` to version 0.7
+- Update `embassy-time` to version 0.4
+- Update `embassy-net` to version 0.6
+- Update `esp-hal` to version 0.23
+- Update `esp-hal-embassy` to version 0.6
+- Update `esp-backtrace` to version 0.15
+- Update `esp-println` to version 0.13
+- Update `esp-wifi` to version 0.12
+- Enable feature `critical-section` on crate `esp-println`
+- Update `rand_core` to version 0.9
+
+
 ## [0.5.0] - 2024-12-22
 
 ### Changed
@@ -81,3 +99,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.3.0]: https://gitlab.com/claudiomattera/esp32c3-embassy/-/tags/0.3.0
 [0.4.0]: https://gitlab.com/claudiomattera/esp32c3-embassy/-/tags/0.4.0
 [0.5.0]: https://gitlab.com/claudiomattera/esp32c3-embassy/-/tags/0.5.0
+[0.6.0]: https://gitlab.com/claudiomattera/esp32c3-embassy/-/tags/0.6.0